+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Fixed
 
 - fix help message of ``prompt`` command in ``prompt`` module
+- fix return value of ``String#lastIndexOf`` when specified empty string
+    - now return haystack size
 
 ## [0.32.0] - 2023-12-31